auth-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic[email]==2.9.2
python-jose[cryptography]==3.3.0
bcrypt==4.1.2
python-multipart==0.0.12
alembic==1.14.0
python-dotenv==1.0.1
email-validator==2.2.0
slowapi==0.1.9

